renamed TourVenueManager to TourManager across the board
